requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.0
requests==2.32.4
httpx==0.28.1
openai==1.90.0
openai-agents==0.0.19
asyncpg==0.30.0
beautifulsoup4==4.13.4
lxml==5.4.0
duckduckgo-search==8.0.4
rich==14.0.0
tqdm==4.67.1
graphviz==0.21
anyio==4.6.2
